requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic[email]
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
mysql-connector-python==8.2.0
cryptography==41.0.3
python-dotenv==1.0.0 